cleanup and rename of side menu task provider
